+++ count/count_digit.c
@@ -6,13 +6,13 @@
 /*   By: yoav <yoav@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/20 13:59:32 by yoav              #+#    #+#             */
-/*   Updated: 2022/06/22 11:30:12 by yoav             ###   ########.fr       */
+/*   Updated: 2022/06/23 16:39:51 by yoav             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-size_t	count_digit_by_base(long n, long base)
+size_t	count_digit_by_base(long n, size_t base)
 {
 	size_t	ret;
